Remove floating point calculations from LMS7002M embeded code

+++ embedded/lms7002m/lms7002m_context.h
@@ -6,7 +6,7 @@
 typedef struct lms7002m_context {
     lms7002m_hooks hooks;
 
-    float reference_clock_hz; ///< Common reference clock for CGEN, SXR, SXT
+    uint32_t reference_clock_hz; ///< Common reference clock for CGEN, SXR, SXT
 } lms7002m_context;
 
 #endif // LIME_LMS7002M_CONTEXT_H
